Ft 20190110 biz invoice

See merge request wanhutong/wanhugou_bg!546

+++ src/main/java/com/wanhutong/backend/modules/biz/web/inventory/BizSkuTransferController.java
@@ -387,7 +387,7 @@ public class BizSkuTransferController extends BaseController {
 //				"inspectRemark":inspectRemark,
 			bizInvoice.setRemarks(bizInvoiceJson.getString("inspectRemark"));
 //				"collLocate":collLocate,
-			bizInvoice.setCollLocate(bizInvoiceJson.getByte("collLocate"));
+			bizInvoice.setCollLocate(bizInvoiceJson.getInteger("collLocate"));
 //				"sendDate":sendDate,
 			bizInvoice.setSendDate(simpleDateFormat.parse(bizInvoiceJson.getString("sendDate")));
 //				"settlementStatus":settlementStatus
